pyproject.toml
[project]
name = "photoping"
version = "0.1.0"
description = "Send random photos from Mac Photos library via email"
requires-python = ">=3.12"
dependencies = [
    "osxphotos",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.14",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
